api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentaion-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.110",
    "sqlalchemy[asyncio]>=2.0",
    "alembic",
    "asyncpg",
    "bcrypt",
    "python-dotenv",
    "httpx",
    "supabase",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentaion_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
